+++ R/waiter.R
@@ -4,14 +4,17 @@
 #' 
 #' @param html HTML content of waiter, generally a spinner, see \code{\link{spinners}}.
 #' @param color Background color of loading screen.
-#' @param logo Path to logo to display.
+#' @param logo Path to logo to display. Deprecated.
 #' @param image Path to background image.
+#' @param fadeout Use a fade out effect when the screen is removed.
+#' Can be a boolean, or a numeric indicating the number of 
+#' milliseconds the effect should take. 
 #' @param id Id of element to hide or element on which to show waiter over.
 #' @param hide_on_render Set to \code{TRUE} to automatically hide the waiter
 #' when the plot in \code{id} is drawn. Note the latter will only work with
 #' shiny plots, tables, htmlwidgets, etc. but will not work with arbitrary
 #' elements.
-#' @param spinners Spinners to include. By default all the CSS files for 
+#' @param spinners Deprecated argument. Spinners to include. By default all the CSS files for 
 #' all spinners are included you can customise this only that which you 
 #' need in order to reduce the amount of CSS that needs to be loaded and
 #' improve page loading speed. There are 7 spinner kits. The spinner kit
@@ -36,8 +39,8 @@
 #' library(shiny)
 #' 
 #' ui <- fluidPage(
-#'   use_waiter(), # dependencies
-#'   waiter_show_on_load(spin_fading_circles()), # shows before anything else 
+#'   useWaiter(), # dependencies
+#'   waiterShowOnLoad(spin_fading_circles()), # shows before anything else 
 #'   actionButton("show", "Show loading for 5 seconds")
 #' )
 #' 
@@ -61,110 +64,46 @@
 #' @import shiny
 #' @name waiter
 #' @export
-use_waiter <- function(spinners = 1:7, include_js = TRUE){
+use_waiter <- function(spinners = NULL, include_js = TRUE){
+  # to deprecate
+  useWaiter(spinners, include_js)
+}
+
+#' @export 
+#' @rdname waiter
+useWaiter <- function(spinners = NULL, include_js = TRUE){
 
   if(!isTRUE(include_js))
-    warning("include_js argument is deprecated, it is no longer needed")
-
-  # must haves
-  header <- tags$head(
-    tags$link(
-      href = "waiter-assets/waiter/waiter.css",
-      rel="stylesheet",
-      type="text/css"
+    .Deprecated(
+      "", 
+      package = "waiter", 
+      msg = "include_js argument is deprecated, it is no longer needed"
     )
+
+  if(!is.null(spinners)){
+    .Deprecated(
+      "",
+      package = "waiter",
+      msg = "spinners argument is no longer in use"
+    )
+  }
+  
+  htmltools::htmlDependency(
+    name = "waiter",
+    version = utils::packageVersion("waiter"),
+    src = "packer",
+    package = "waiter",
+    script = "waiter.js"
   )
-
-  # spinner kits
-  if(1 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/spinkit.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(2 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/css-spinners.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(3 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/devloop.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(4 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/spinners.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(5 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/spinbolt.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(6 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/loaders.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  if(7 %in% spinners)
-    header <- shiny::tagAppendChildren(
-      header,
-      tags$link(
-        href = "waiter-assets/waiter/custom.css",
-        rel="stylesheet",
-        type="text/css"
-      )
-    )
-
-  # add js
-  header <- shiny::tagAppendChildren(
-    header,
-    tags$script(
-      src = "waiter-assets/waiter/waiter.js"
-    ),
-    tags$script(
-      src = "waiter-assets/waiter/custom.js"
-    )
-  )
-
-  # singleton it
-  singleton(header)
 
 }
 
 #' @rdname waiter
 #' @export
-waiter_use <- use_waiter
+waiter_use <- function(spinners = 1:7, include_js = TRUE){
+  # to deprecate
+  useWaiter(spinners, include_js)
+}
 
 #' @rdname waiter
 #' @export
@@ -174,7 +113,16 @@ waiter_show <- function(
   color = "#333e48", 
   logo = "", 
   image = "",
-  hide_on_render = !is.null(id)){
+  hide_on_render = !is.null(id)
+){
+
+  if(logo != "") {
+    .Deprecated(
+      "html", 
+      package = "waiter", 
+      msg = "The logo argument is deprecated, use html instead"
+    )
+  }
   
   html <- as.character(html)
   html <- gsub("\n", "", html)
@@ -186,34 +134,52 @@ waiter_show <- function(
     id = id,
     html = html,
     color = color,
-    logo = logo,
     image = image,
     hide_on_render = hide_on_render
   )
   session <- shiny::getDefaultReactiveDomain()
-  .check_session(session)
   session$sendCustomMessage("waiter-show", opts)
 }
-
 
 #' @rdname waiter
 #' @export
 waiter_show_on_load <- function(
   html = spin_1(), 
   color = "#333e48",
-  image = ""
+  image = "",
+  logo = ""
 ){
+  # to deprecate
+  waiterShowOnLoad(html, color, image, logo)
+}
+
+#' @rdname waiter
+#' @export
+waiterShowOnLoad <- function(
+  html = spin_1(), 
+  color = "#333e48",
+  image = "",
+  logo = ""
+){
+  
+  if(logo != "") {
+    .Deprecated(
+      "html", 
+      package = "waiter", 
+      msg = "The logo argument is deprecated, use html instead"
+    )
+  }
   
   html <- as.character(html)
   html <- gsub("\n", "", html)
 
   show <- sprintf(
-    "show_waiter(
-      null,
-      html = '%s', 
-      color = '%s',
-      image = '%s'
-    );",
+    "waiter.show({
+      id: null,
+      html: '%s', 
+      color: '%s',
+      image: '%s'
+    });",
     html, color, image
   )
 
@@ -226,27 +192,53 @@ waiter_show_on_load <- function(
 waiter_preloader <- function(
   html = spin_1(), 
   color = "#333e48",
-  image = ""
+  image = "",
+  fadeout = FALSE,
+  logo = ""
 ){
+  # to deprecate
+  waiterPreloader(html, color, image, fadeout, logo)
+}
+
+#' @rdname waiter
+#' @export
+waiterPreloader <- function(
+  html = spin_1(), 
+  color = "#333e48",
+  image = "",
+  fadeout = FALSE,
+  logo = ""
+){
+  
+  if(logo != "") {
+    .Deprecated(
+      "html", 
+      package = "waiter", 
+      msg = "The logo argument is deprecated, use html instead"
+    )
+  }
   
   html <- as.character(html)
   html <- gsub("\n", "", html)
+  
+  fadeout <- ifelse(is.logical(fadeout), tolower(fadeout), fadeout)
 
   show <- sprintf(
-    "show_waiter(
-      null,
-      html = '%s', 
-      color = '%s',
-      image = '%s',
-    );",
-    html, color, image
+    "waiter.show({
+      id: null,
+      html: '%s', 
+      color: '%s',
+      image: '%s',
+      fadeOut: %s
+    });",
+    html, color, image, fadeout
   )
 
   hide <- paste0(
     "window.ran = false;",
     "$(document).on('shiny:idle', function(event){
       if(!window.ran)
-        hide_waiter(id = null);
+        waiter.hideWaiter(id = null);
 
       window.ran = true;
     });"
@@ -267,13 +259,20 @@ waiter_preloader <- function(
 #' @rdname waiter
 #' @export
 waiter_hide_on_render <- function(id){
+  # to deprecate
+  waiterHideOnRender(id)
+}
+
+#' @rdname waiter
+#' @export
+waiterHideOnRender <- function(id){
   if(missing(id))
     stop("Missing id", call. = FALSE)
   
   script <- sprintf(
     "$(document).on('shiny:value', function(event) {
       if(event.name == '%s'){
-        hide_waiter(null);
+        waiter.hideWaiter(null);
       }
     });",
     id
@@ -288,23 +287,36 @@ waiter_hide_on_render <- function(id){
 
 #' @rdname waiter
 #' @export
-waiter_on_busy <- function(html = spin_1(), color = "#333e48", logo = "", image = ""){
+waiter_on_busy <- function(html = spin_1(), color = "#333e48", logo = "", image = "", fadeout = FALSE){
+  # to deprecate
+  waiterOnBusy(html, color, logo, image, fadeout)
+}
+
+#' @rdname waiter
+#' @export
+waiterOnBusy <- function(html = spin_1(), color = "#333e48", logo = "", image = "", fadeout = FALSE){
+
+  if(logo != "")
+    .Deprecated("html", "waiter", "The logo argument is deprecated")
 
   html <- as.character(html)
   html <- gsub("\n", "", html)
 
+  fadeout <- ifelse(is.logical(fadeout), tolower(fadeout), fadeout)
+
   script <- paste0(
     "$(document).on('shiny:busy', function(event) {
-      show_waiter(
-        id = null,
-        html = '", html, "', 
-        color = '", color, "',
-        image = '", image, "'
-      );
+      waiter.show({
+        id: null,
+        html: '", html, "', 
+        color: '", color, "',
+        image: '", image, "',
+        fadeOut: ", fadeout, "
+      });
     });
     
     $(document).on('shiny:idle', function(event) {
-      hide_waiter(null);
+      waiter.hideWaiter(null);
     });"
   )
 
@@ -315,7 +327,6 @@ waiter_on_busy <- function(html = spin_1(), color = "#333e48", logo = "", image 
 #' @export
 waiter_hide <- function(id = NULL){
   session <- shiny::getDefaultReactiveDomain()
-  .check_session(session)
   session$sendCustomMessage("waiter-hide", list(id = id))
 }
 
@@ -334,7 +345,6 @@ waiter_update <- function(id = NULL, html = NULL){
   html <- as.character(html)
   html <- gsub("\n", "", html)
   session <- shiny::getDefaultReactiveDomain()
-  .check_session(session)
   session$sendCustomMessage("waiter-update", list(html = html, id = id))
 }
 
@@ -358,7 +368,10 @@ Waiter <- R6::R6Class(
 #' @param html HTML content of waiter, generally a spinner, see \code{\link{spinners}} or a list of the latter.
 #' @param color Background color of loading screen.
 #' @param image Path to background image of loading screen.
-#' @param logo Logo to display.
+#' @param fadeout Use a fade out effect when the screen is removed.
+#' Can be a boolean, or a numeric indicating the number of 
+#' milliseconds the effect should take. 
+#' @param logo Logo to display. Deprecated.
 #' @param id Id, or vector of ids, of element on which to overlay the waiter, if \code{NULL} the waiter is
 #' applied to the entire body.
 #' @param hide_on_render Set to \code{TRUE} to automatically hide the waiter
@@ -371,13 +384,15 @@ Waiter <- R6::R6Class(
 #' @examples
 #' \dontrun{Waiter$new()}
     initialize = function(id = NULL, html = NULL, color = NULL, logo = NULL, 
-      image = "", hide_on_render = !is.null(id), hide_on_error = !is.null(id),
+      image = "", fadeout = FALSE, hide_on_render = !is.null(id), hide_on_error = !is.null(id),
       hide_on_silent_error = !is.null(id)){
+      
+      if(!is.null(logo) && logo != "")
+        .Deprecated("html", "waiter", "The logo argument is deprecated")
 
       # get theme
       html <- .theme_or_value(html, "WAITER_HTML")
       color <- .theme_or_value(color, "WAITER_COLOR")
-      logo <- .theme_or_value(logo, "WAITER_LOGO")
       image <- .theme_or_value(image, "WAITER_IMAGE")
       
       # process inputs
@@ -407,7 +422,7 @@ Waiter <- R6::R6Class(
       private$.html <- html
       private$.color <- color
       private$.image <- image
-      private$.logo <- logo
+      private$.fadeout <- fadeout
       private$.hide_on_render <- hide_on_render
       private$.hide_on_silent_error <- hide_on_silent_error
       private$.hide_on_error <- hide_on_error
@@ -421,11 +436,12 @@ Waiter <- R6::R6Class(
           id = private$.id[[i]],
           html = private$.html[[i]],
           color = private$.color,
-          logo = private$.logo,
           image = private$.image,
-          hide_on_render = private$.hide_on_render,
-          hide_on_silent_error = private$.hide_on_silent_error,
-          hide_on_error = private$.hide_on_error
+          hideOnRender = private$.hide_on_render,
+          hideOnSilentError = private$.hide_on_silent_error,
+          hideOnError = private$.hide_on_error,
+          fadeOut = private$.fadeout,
+          blur = private$.blur
         )
         private$.session$sendCustomMessage("waiter-show", opts)
       }
@@ -465,22 +481,59 @@ Waiter <- R6::R6Class(
       if(!is.null(private$.id))
 		    cat("A waiter on", paste0(private$.id, collapse = ","), "\n")
       else
-        cat("A waiter\n")
+        cat("A full-screen waiter\n")
 		}
+  ),
+  active = list(
+#' @field fadeout Set or get the fade out
+    fadeout = function(value) {
+      if(missing(value))
+        return(private$.fadeout)
+      
+      private$.fadeout <- value
+    },
+#' @field color Set or get the background color
+    color = function(value) {
+      if(missing(value))
+        return(private$.color)
+      
+      private$.color <- value
+    },
+#' @field image Set of get the background image
+    image = function(value) {
+      if(missing(value))
+        return(private$.image)
+      
+      private$.image <- value
+    },
+#' @field session Set or get the shiny session
+    session = function(value) {
+      if(missing(value))
+        return(private$.session)
+      
+      private$.session <- value
+    },
+#' @field html Set or get the html content
+    html = function(value) {
+      if(missing(value))
+        return(private$.html)
+      
+      private$.html <- value
+    }
   ),
   private = list(
     .html = list(),
     .color = "#333e48",
-    .logo = "",
     .image = "",
     .id = NULL,
+    .blur = NULL,
     .session = NULL,
+    .fadeout = FALSE,
     .hide_on_render = FALSE,
     .hide_on_silent_error = FALSE,
     .hide_on_error = FALSE,
 		get_session = function(){
 			private$.session <- shiny::getDefaultReactiveDomain()
-			.check_session(private$.session)
 		}
   )
 )
@@ -495,10 +548,13 @@ Waiter <- R6::R6Class(
 #' @name waiterTheme
 #' @export
 waiter_set_theme <- function(html = spin_1(), color = "#333e48", logo = "", image = ""){
+  
+  if(logo != "")
+    .Deprecated("html", "waiter", "The logo argument is deprecated")
+
   options(
     WAITER_HTML = html,
     WAITER_COLOR = color,
-    WAITER_LOGO = logo,
     WAITER_IMAGE = image
   )
   invisible()
@@ -510,7 +566,6 @@ waiter_get_theme <- function(){
   list(
     html = .get_html(),
     color = .get_color(),
-    logo = .get_logo(),
     image = .get_image()
   )
 }
@@ -521,7 +576,6 @@ waiter_unset_theme <- function(){
   options(
     WAITER_HTML = NULL,
     WAITER_COLOR = NULL,
-    WAITER_LOGO = NULL,
     WAITER_IMAGE = NULL
   )
   invisible()
@@ -543,4 +597,205 @@ transparent <- function(alpha = 0){
     stop("`alpha` must be between 0 and 1", call. = FALSE)
   
   invisible(paste0("rgba(255,255,255,", alpha, ")"))
+}
+
+#' Trigger Waiter
+#' 
+#' A a trigger to a waiting screen from the UI.
+#' 
+#' @param el Element that triggers the waiter.
+#' @param on The event that triggers the waiter.
+#' @param hide_on_error,hide_on_silent_error Whether to hide the waiter when the underlying element throws an error.
+#' Silent error are thrown by \link[shiny]{req} and  \link[shiny]{validate}.
+#' @inheritParams waiter
+#' 
+#' @examples
+#' library(shiny)
+#' library(waiter)
+#' 
+#' ui <- fluidPage(
+#'  useWaiter(),
+#'  triggerWaiter(
+#'    actionButton(
+#'      "generate",
+#'      "Generate Plot"
+#'    )
+#'  ),
+#'  plotOutput("plot")
+#' )
+#' 
+#' server <- function(input, output){
+#'  output$plot <- renderPlot({
+#'    input$generate
+#'    Sys.sleep(3)
+#'    plot(runif(50))
+#'  })
+#' }
+#' 
+#' if(interactive())
+#'  shinyApp(ui, server)
+#' 
+#' @export
+triggerWaiter <- function(
+  el,
+  id = NULL,
+  html = NULL, 
+  color = NULL, 
+  image = "", 
+  fadeout = FALSE,
+  on = "click", 
+  hide_on_render = !is.null(id), 
+  hide_on_error = !is.null(id),
+  hide_on_silent_error = !is.null(id) 
+){
+
+  # get/set id of trigger
+  el_id <- el$attr$id
+
+  if(is.null(el_id)){
+    el_id <- .random_name()
+    shiny::tagAppendAttributes(el, id = el_id)
+  }
+  
+  # html
+  html <- .theme_or_value(html, "WAITER_HTML")
+  html <- as.character(html)
+  html <- gsub("\n", "", html)
+
+  color <- .theme_or_value(color, "WAITER_COLOR")
+  fadeout <- ifelse(is.logical(fadeout), tolower(fadeout), fadeout)
+
+  if(!is.null(id))
+    id <- sprintf("'%s'", id)
+
+  if(is.null(id))
+    id <- "null"
+
+  script <- sprintf(
+    "$('#%s').on('%s', function(event){
+      waiter.show({
+        id: %s,
+        html: '%s', 
+        color: '%s', 
+        hideOnRender: %s, 
+        hideOnError: %s, 
+        hideOnSilentError: %s, 
+        image: '%s',
+        fadeOut: %s
+      });
+    })",
+    el_id,
+    on,
+    id, 
+    html,
+    color,
+    tolower(hide_on_render),
+    tolower(hide_on_silent_error),
+    tolower(hide_on_silent_error),
+    image,
+    fadeout
+  )
+
+  shiny::tagList(
+    el,
+    HTML(sprintf("<script>%s</script>", script))
+  )
+}
+
+#' Automatic Waiter
+#' 
+#' This function allows easily adding
+#' waiters to dynamically rendered Shiny content where
+#' "dynamic" means `render*` and `*output` function pair.
+#' 
+#' This will display the waiter when the element is being 
+#' recalculated and hide it when it receives new data.
+#' 
+#' @param id Vector of ids of elements to overlay the waiter.
+#' If `NULL` then the loading screens are applied to all
+#' elements.
+#' @inheritParams waiter
+#' 
+#' @examples
+#' library(shiny)
+#' library(waiter)
+#' 
+#' ui <- fluidPage(
+#' 	autoWaiter(),
+#' 	actionButton(
+#' 		"trigger",
+#' 		"Render"
+#' 	),
+#' 	plotOutput("plot"),
+#' 	plotOutput("dom")
+#' )
+#' 
+#' server <- function(input, output){
+#' 	output$plot <- renderPlot({
+#' 		input$trigger
+#' 		Sys.sleep(3)
+#' 		plot(cars)
+#' 	})
+#' 
+#' 	output$dom <- renderPlot({
+#' 		input$trigger
+#' 		Sys.sleep(5)
+#' 		plot(runif(100))
+#' 	})
+#' }
+#' 
+#' if(interactive())
+#'  shinyApp(ui, server)
+#' 
+#' @export
+autoWaiter <- function(
+  id = NULL,
+  html = NULL, 
+  color = NULL, 
+  image = "", 
+  fadeout = FALSE
+){
+
+  # html
+  html <- .theme_or_value(html, "WAITER_HTML")
+  html <- as.character(html)
+  html <- gsub("\n", "", html)
+
+  color <- .theme_or_value(color, "WAITER_COLOR")
+  fadeout <- ifelse(is.logical(fadeout), tolower(fadeout), fadeout)
+
+  auto <- ifelse(is.null(id), "true", "false")
+  ids <- paste0("['", paste0(id, collapse = "','"), "']")
+
+  script <- sprintf(
+    "let auto = %s;
+    $(document).on('shiny:recalculating', function(event) {
+
+      if(!auto)
+        if(!%s.includes(event.target.id))
+          return ;
+
+      waiter.show({
+        id: event.target.id,
+        html: '%s', 
+        color: '%s', 
+        hideOnRender: true, 
+        hideOnError: true, 
+        hideOnSilentError: true, 
+        image: '%s',
+        fadeOut: %s
+      });
+    });",
+    auto,
+    ids,
+    html,
+    color,
+    image,
+    fadeout
+  )
+
+  shiny::tagList(
+    useWaiter(),
+    HTML(sprintf("<script>%s</script>", script))
+  )
 }